[FIX]: Fixed bug on changing product detail, FeatureContainer.jsx -similarProducts mode- works now with local state instead of redux.
