+++ src/redux/action-type.js
@@ -6,7 +6,6 @@ export const GET_PRODUCT_BY_NAME = 'GET_PRODUCT_BY_NAME'
 export const GET_PRODUCT_BY_ID = 'GET_PRODUCT_BY_ID'
 export const GET_ALL_PRODUCTS_NO_FILTER = 'GET_ALL_PRODUCTS_NO_FILTER'
 export const CLEAN_PRODUCT_DETAIL = 'CLEAN_PRODUCT_DETAIL'
-export const GET_SIMILAR_PRODUCTS = 'GET_SIMILAR_PRODUCTS'
 
 // CATEGORIES
 export const GET_ALL_CATEGORIES = 'GET_ALL_CATEGORIES'
