+++ src/services/api.js
@@ -46,3 +46,36 @@ export const get_all_products_filtered = async (query) => {
         throw error
     }
 }
+
+export const get_similar_products = async ({ currentId, limit, category, color }) => {
+    try {
+        // Construir la URL incial con un parámetro común.
+        let url = ['products?active=true']
+        // Concatenar paramétro "page".
+        url.push(`&page=1`)
+        // Concatenar el resto de paramétros: category.
+        if (category) url.push(`&category=${category}`)
+        if (color) url.push(`&color=${color}`)
+        // Juntar el array en una sola string
+        const query = url.join('')
+
+        // Hacer la petición de la api con la query creada para obtener los productos filtrados.
+        const data = await get_all_products_filtered(query)
+        const similarProducts = data.results.rows.filter((product) => product.idProduct !== currentId).slice(0, limit)
+
+        if (similarProducts.length < 2) {
+            url = ['products?active=true'];
+            if (category) url.push(`&category=${category}`)
+            const query = url.join('')
+            // Hacer la petición de la api con la query creada para obtener los productos filtrados.
+            const data = await get_all_products_filtered(query)
+            const similarProducts = data.results.rows.filter((product) => product.idProduct !== currentId).slice(0, limit)
+            return similarProducts;
+        } else {
+            return similarProducts;
+        }
+    } catch (error) {
+        console.log('Error trying to dispatch getSimilarProducts: ' + error)
+        return [];
+    }
+}
